requirements.txt
Django==1.4
South==0.7.4
distribute==0.6.24
django-debug-toolbar==0.9.4
facebook-sdk==0.3.0
gunicorn==0.14.2
psycopg2==2.4.5
tweepy==1.9
wsgiref==0.1.2
